test: 카드 소유자 이름 입력 필드 테스트 추가
